Struct rustpi_io::gpio::GPIO

pub struct GPIO { /* fields omitted */ }

A software representation for a GPIO pin of the raspberry pi.

Methods

impl GPIO

pub fn gpio_number(&self) -> u8

Returns the configured pin of the GPIO struct.

pub fn current_mode(&self) -> GPIOMode

Returns the configured mode of the GPIO struct.

pub fn set_mode(&mut self, mode: GPIOMode) -> Result<&mut Self>

Changes the mode of the pin and writes the corresponding value to the fitting direction file

pub fn new(gpio: u8, mode: GPIOMode) -> Result<Self>

Initializes the gpio. Exports the pin with the /sys/class/gpio/export file and calls the set_mode() function with the given mode. Returns an Error if the gpio was already exported earlier (inside or outside of the application)

pub fn value(&self) -> Result<GPIOData>

Reads the current value of the pin in both Read and Write mode. Returns an Error if a value other than "1" or "0" is read

pub fn set(&self, data: GPIOData) -> Result<()>

Sets the value of the gpio to HIGH or LOW Returns an Error if the GPIO::Mode is not Write