Module rustpi_io::serial

Structs

SerialPi
Interface for the spi bus of the Raspberry Pi.

Enums

ComMode

Device
Correspond to the SPI Chip Enable Pins on the raspberry pi.

Speed
From https://www.raspberrypi.org/documentation/hardware/raspberrypi/spi/README.md#driver
Possible Speeds:
125.0 MHz
62.5 MHz
31.2 MHz
15.6 MHz
7.8 MHz
3.9 MHz
1953 kHz
976 kHz
488 kHz
244 kHz
122 kHz
61 kHz
30.5 kHz
15.2 kHz
7629 Hz

SpiMode
The most common spi modes regulating the clock phase and polarity.
Mode 0 seems to be the most used one and is set as default.
See https://en.wikipedia.org/wiki/Serial_Peripheral_Interface_Bus#Clock_polarity_and_phase for an explanation