Enum rustpi_io::serial::SpiMode

pub enum SpiMode {
    Mode0,
    Mode1,
    Mode2,
    Mode3,
}

The most common spi modes regulating the clock phase and polarity.
Mode 0 seems to be the most used one and is set as default.
See https://en.wikipedia.org/wiki/Serial_Peripheral_Interface_Bus#Clock_polarity_and_phase for an explanation

Variants
Mode0Mode1Mode2Mode3