Struct rustpi_io::serial::SerialPi

pub struct SerialPi {
    pub com_mode: ComMode,
    // some fields omitted
}

Interface for the spi bus of the Raspberry Pi.

A SerialPi spi struct can be opened to access the serial interface of the Raspberry Pi as Master. It will initialize the necessary components for the communication. A device, a bus clock, a spi mode and a communication mode have to be configured.

The device correspond to the Chip Enable Pins of the raspberry pi. Look at the pin layout of of your raspberry pi to determine which pin is associated with which CE device.

The bus clock determines the communication speed. Use a speed that your slave device can handle.

The spi mode sets clock polarity and clock phase of the transmission. The configuration has to match with the slave device. Mode 0 is said to be the most common. For an explanation you can go to the wikipedia spi article.

The communication mode can be Full or HalfDuplex. In full duplex mode, for every byte that is send, a byte will be received. In half duplex mode the bytes that are received during a transmission are ignored.

In full duplex mode, the data that was received with write will be buffered internally. Calls to the Read and BufRead trait implementation will first read out the buffered content and will issue an actual spi read if the buffer is exhausted. The buffer will be reallocated if the last read bytes do not fit. It will preserve the capacity after that until a resize is issued with try_shrink_to.

In half duplex mode the buffer will not be filled when calling write, but it will be consumed when calling read.

Fields
com_mode: ComMode
Methods

impl SerialPi

pub fn new(
device: Device,
speed: Speed,
spi_mode: SpiMode,
communication_mode: ComMode
) -> Result<SerialPi>

Calls with_capacity with a buffer size of 1000 bytes.

pub fn with_capacity(
device: Device,
speed: Speed,
spi_mode: SpiMode,
communication_mode: ComMode,
buffer_capacity: usize
) -> Result<SerialPi>

Creates a serial wrapper for the raspberry pi. Also sets the buffers capacity.

Note: The SpiMode and the ComMode have a default value.

Errors

Can return an error if the spi device can't be opened. It might be already in use or the raspberry is not configured correctly. Check the documentation of the raspberry pi in this case.

pub fn buffer_capacity(&self) -> usize

Returns the current capacity of the internal buffer. To change it call try_shrink_to. Maybe you have to consume or read some bytes of the buffer first.

pub fn try_shrink_to(&mut self, desired_capacity: usize) -> usize

Shrinks the internal buffer to fit its length. If the new capacity is les then the desired, the capacity is extended to match the desired. Already read bytes are not dropped. To do that call consume. The actual capacity might be greater than the desired.